OBPIH-3264 Ensure that when transactions are created or updated that the entire database transaction is flushed or rolled back together
